Revert "reformatted codebase and fixed single commands"
