+++ src/main/java/com/ryanmichela/sshd/ConsoleCommandFactory.java
@@ -1,6 +1,5 @@
 package com.ryanmichela.sshd;
 
-import com.ryanmichela.sshd.implementations.SSHDCommandSender;
 import org.apache.sshd.server.command.Command;
 import org.apache.sshd.server.command.CommandFactory;
 import org.apache.sshd.server.channel.ChannelSession;
@@ -24,4 +23,59 @@ public class ConsoleCommandFactory implements CommandFactory
         return new ConsoleCommand(command);
     }
 
+    public class ConsoleCommand implements Command 
+    {
+        private String command;
+        private InputStream in;
+        private OutputStream out;
+        private OutputStream err;
+        private ExitCallback callback;
+
+        public ConsoleCommand(String command) 
+        {
+            this.command = command;
+        }
+
+        public void setInputStream(InputStream in) 
+        {
+            this.in = in;
+        }
+
+        public void setOutputStream(OutputStream out) 
+        {
+            this.out = out;
+        }
+
+        public void setErrorStream(OutputStream err) 
+        {
+            this.err = err;
+        }
+
+        public void setExitCallback(ExitCallback callback) 
+        {
+            this.callback = callback;
+        }
+
+        @Override
+        public void start(ChannelSession cs, Environment environment) throws IOException 
+        {
+            try 
+            {
+                SshdPlugin.instance.getLogger()
+                        .info("[U: " + environment.getEnv().get(Environment.ENV_USER) + "] " + command);
+                Bukkit.dispatchCommand(Bukkit.getConsoleSender(), command);
+            } 
+            catch (Exception e) 
+            {
+                SshdPlugin.instance.getLogger().severe("Error processing command from SSH -" + e.getMessage());
+            } 
+            finally 
+            {
+                callback.onExit(0);
+            }
+        }
+
+        @Override
+        public void destroy(ChannelSession cn) {}
+	}
 }
